requirements.txt
discord
python-dotenv
art
colorama
